+++ CHANGELOG.md
@@ -10,7 +10,11 @@
 **Internal**:
 
 - Add sampling + tagging by event platform and transaction op. Some (unused) tagging rules from 22.4.0 have been renamed. ([#1231](https://github.com/getsentry/relay/pull/1231))
+- Refactor aggregation error, recover from errors more gracefully. ([#1240](https://github.com/getsentry/relay/pull/1240))
+- Remove/reject nul-bytes from metric strings. ([#1235](https://github.com/getsentry/relay/pull/1235))
+- Remove the unused "internal" data category. ([#1245](https://github.com/getsentry/relay/pull/1245))
 - Add ReplayRecording ItemType. ([#1236](https://github.com/getsentry/relay/pull/1236))
+
 
 ## 22.4.0